Fix all import paths after services modularization
- Updated all @/services/ imports to subdirectories (ai/, browser/, features/)
- Updated all ~/services/ imports to subdirectories
- Build now succeeds for Chrome and Firefox
