+++ src/components/Layouts/Header.tsx
@@ -16,7 +16,7 @@ import { SelectedKnowledge } from "../Option/Knowledge/SelectedKnowledge"
 import { ModelSelect } from "../Common/ModelSelect"
 import { PromptSelect } from "../Common/PromptSelect"
 import { useQuery } from "@tanstack/react-query"
-import { fetchChatModels } from "~/services/ollama"
+import { fetchChatModels } from "~/services/ai/ollama"
 import { useMessageOption } from "~/hooks/useMessageOption"
 import { Avatar, Select, Tooltip } from "antd"
 import { getAllPrompts } from "@/db/dexie/helpers"
