+++ src/components/Option/Knowledge/AddKnowledge.tsx
@@ -1,6 +1,6 @@
 import { createKnowledge } from "@/db/dexie/knowledge"
 import { Source } from "@/db/knowledge"
-import { defaultEmbeddingModelForRag } from "@/services/ollama"
+import { defaultEmbeddingModelForRag } from "@/services/ai/ollama"
 import { convertTextToSource, convertToSource } from "@/utils/to-source"
 import { useMutation } from "@tanstack/react-query"
 import { Modal, Form, Input, Upload, message, Tabs, Select } from "antd"
