+++ src/components/Option/Models/AddUpdateModelSettings.tsx
@@ -1,4 +1,4 @@
-import { getModelSettings, setModelSettings } from "@/services/model-settings"
+import { getModelSettings, setModelSettings } from "@/services/ai/model-settings"
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
 import {
   Collapse,
