+++ src/components/Option/Settings/model-settings.tsx
@@ -1,6 +1,6 @@
 import { BetaTag } from "@/components/Common/BetaTag"
 import { SaveButton } from "@/components/Common/SaveButton"
-import { getAllModelSettings, setModelSetting } from "@/services/model-settings"
+import { getAllModelSettings, setModelSetting } from "@/services/ai/model-settings"
 import { useQuery, useQueryClient } from "@tanstack/react-query"
 import { Form, Skeleton, Input, InputNumber, Collapse, Switch } from "antd"
 import React from "react"
