+++ src/components/Option/Settings/ollama.tsx
@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query"
 import { Collapse, Skeleton, Switch } from "antd"
 import { useState } from "react"
 import { SaveButton } from "~/components/Common/SaveButton"
-import { getOllamaURL, setOllamaURL as saveOllamaURL } from "~/services/ollama"
+import { getOllamaURL, setOllamaURL as saveOllamaURL } from "~/services/ai/ollama"
 import { Trans, useTranslation } from "react-i18next"
 import { AdvanceOllamaSettings } from "@/components/Common/Settings/AdvanceOllamaSettings"
 import { ModelSettings } from "./model-settings"
