+++ src/components/Option/Settings/rag.tsx
@@ -9,10 +9,10 @@ import {
   defaultSsplttingSeparator,
   getEmbeddingModels,
   saveForRag
-} from "~/services/ollama"
+} from "~/services/ai/ollama"
 import { SettingPrompt } from "./prompt"
 import { useTranslation } from "react-i18next"
-import { getNoOfRetrievedDocs, getTotalFilePerKB } from "@/services/app"
+import { getNoOfRetrievedDocs, getTotalFilePerKB } from "@/services/features/app"
 import { SidepanelRag } from "./sidepanel-rag"
 import { ProviderIcons } from "@/components/Common/ProviderIcon"
 import { SettingTitle } from "./title"
