+++ src/components/Option/Settings/title.tsx
@@ -9,10 +9,10 @@ import {
   titleGenerationModel,
   DEFAULT_TITLE_GEN_PROMPT,
   setTitleGenerationModel
-} from "~/services/title"
+} from "~/services/features/title"
 import { ProviderIcons } from "@/components/Common/ProviderIcon"
 import { useStorage } from "@plasmohq/storage/hook"
-import { fetchChatModels } from "@/services/ollama"
+import { fetchChatModels } from "@/services/ai/ollama"
 
 export const SettingTitle = () => {
   const [form] = Form.useForm()
