+++ src/components/Option/Share/index.tsx
@@ -4,7 +4,7 @@ import { Trash2 } from "lucide-react"
 import { Trans, useTranslation } from "react-i18next"
 import { SaveButton } from "~/components/Common/SaveButton"
 import { deleteWebshare, getAllWebshares, getUserId } from "@/db/dexie/helpers"
-import { getPageShareUrl, setPageShareUrl } from "~/services/ollama"
+import { getPageShareUrl, setPageShareUrl } from "~/services/ai/ollama"
 import { verifyPageShareURL } from "~/utils/verify-page-share"
 import { useStorage } from "@plasmohq/storage/hook"
 import fetcher from "@/libs/fetcher"
