+++ src/hooks/chat-modes/documentChatMode.ts
@@ -4,7 +4,7 @@ import {
   getOllamaURL,
   geWebSearchFollowUpPrompt,
   promptForRag
-} from "~/services/ollama"
+} from "~/services/ai/ollama"
 import { type ChatHistory, type Message } from "~/store/option"
 import { addFileToSession, generateID, getSessionFiles } from "@/db/dexie/helpers"
 import { generateHistory } from "@/utils/generate-history"
@@ -18,13 +18,13 @@ import {
 } from "@/libs/reasoning"
 import { getModelNicknameByID } from "@/db/dexie/nickname"
 import { formatDocs } from "@/chain/chat-with-x"
-import { getAllDefaultModelSettings } from "@/services/model-settings"
-import { getNoOfRetrievedDocs } from "@/services/app"
+import { getAllDefaultModelSettings } from "@/services/ai/model-settings"
+import { getNoOfRetrievedDocs } from "@/services/features/app"
 import { pageAssistEmbeddingModel } from "@/models/embedding"
 import { UploadedFile } from "@/db/dexie/types"
 import { getSystemPromptForWeb, isQueryHaveWebsite } from "@/web/web"
 import { PAMemoryVectorStore } from "@/libs/PAMemoryVectorStore"
-import { getMaxContextSize } from "@/services/kb"
+import { getMaxContextSize } from "@/services/features/kb"
 
 export const documentChatMode = async (
   message: string,
