+++ src/web/web.ts
@@ -1,7 +1,7 @@
-import { getWebSearchPrompt } from "~/services/ollama"
+import { getWebSearchPrompt } from "~/services/ai/ollama"
 import { webGoogleSearch } from "./search-engines/google"
 import { webDuckDuckGoSearch } from "./search-engines/duckduckgo"
-import { getIsVisitSpecificWebsite, getSearchProvider, getDomainFilterList, getBlockedDomainList } from "@/services/search"
+import { getIsVisitSpecificWebsite, getSearchProvider, getDomainFilterList, getBlockedDomainList } from "@/services/features/search"
 import { webSogouSearch } from "./search-engines/sogou"
 import { webBraveSearch } from "./search-engines/brave"
 import { getWebsiteFromQuery, processSingleWebsite } from "./website"
